Compile everything, then run
This is a major change in how the bytecode interpreter works.
We now compile all statements first and then run everything in the VM.
This is conceptually much cleaner and in theory also faster. In
practice, most of the time is spent elsewhere, so it hardly matters (at
the moment).

We also store much fewer constants in the VM by adding a new
`ApplyPrefix` operation.

closes #22