Unit representation: store direct representation
This is a major change to how units are represented in Numbat.
Previously, derived units were always defined by their relation to base
units. For example: A foot was represented as 0.3048 meter, even though
it was defined as 12 inch.

With this change, we represent those unit relationships closer to how
they have been defined, i.e. 1 foot is a defined unit with a conversion
factor of 12 and a defining unit of inch.

This is preparatory work that will make further changes to unit
conversion and quantity simplification possible.