Wait for background currency fetch thread to finish before exiting
When numbat is run with a quick-running --expression, it can finish
evaluating the expression before the background currency fetch thread
has finished.

On Windows, this is problematic because the main thread does some
post-main winsock shutdown tasks, which conflict with the background
task that's still trying to fetch exchange rate info from the network.

This commit keeps a reference to the currency fetching thread and
joins on it before exiting.