Always embed modules
This enables embedding of modules into the binary, even in debug mode.
The idea of rust-embed to load files from disk when in debug mode is
neat, but I've tripped over this one too many times. It's just confusing
that things behave differently in debug and release mode. So let's fix
this.