Implement service worker to serve site when offline
This is the first step in making Numbat fully offline-capable and
installable as a PWA. It's a bit brittle right now because it relies on
static assets being manually tracked and added to the service worker
lists.

There's also a hole where the Google Fonts are not explicitly
cached by the service worker; however the browser should be able to fall
back to the system fonts or browser-level cache in this case. I tested
this on my machine and it rendered properly when I was fully offline.

I was not able to test the ECB exchange rate fetching due to the lack of
CORS policy headers on the Numbat proxy, since I was running this at
localhost.