Deduplicate VM constants
(except when defining derived unit constants, as initially identical but
temporary values)

Don't see any real performance changes from this (if anything it's
probably a minor slowdown due to the added dupe check in
`add_constant`). However, it does reduce the number of constants
generated from the prelude from ~730 to ~320, a reduction of 56%. This
in theory allows for larger Numbat programs as it's now harder to hit
the ~65k constants ceiling. Worth it? I don't know :)